# Conflicts:
#	.gitignore
#	src/main/java/nz/ac/canterbury/seng302/homehelper/controller/EditProfileController.java
#	src/main/java/nz/ac/canterbury/seng302/homehelper/entity/User.java
#	src/main/java/nz/ac/canterbury/seng302/homehelper/security/SecurityConfiguration.java
#	src/main/java/nz/ac/canterbury/seng302/homehelper/validation/UserValidation.java

+++ src/main/java/nz/ac/canterbury/seng302/homehelper/controller/RenovationController.java
@@ -8,11 +8,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
@@ -23,6 +22,7 @@ import java.util.List;
  * @author Abhisekh Chand
  */
 @Controller
+@RequestMapping("/renovations")
 public class RenovationController {
 
     private static final Logger logger = LoggerFactory.getLogger(RenovationController.class);
@@ -44,23 +44,17 @@ public class RenovationController {
 
     /**
      * Gets all renovations
-     * @param name optional string to search on renovation name (partial matching)
+     * @param searchQuery optional string to search on renovation name (partial matching)
      * @param model (map-like) representation of results to be used by thymeleaf
      * @return thymeleaf renovationsTemplate
      */
-    @GetMapping("/renovations")
-    public String renovations(@RequestParam(value = "name", required = false) String name, Model model) {
+    @GetMapping
+    public String renovations(@RequestParam(value = "searchQuery", required = false, defaultValue="") String searchQuery, Model model) {
         logger.info("GET renovations");
-        //TODO change cases
         try {
             User user = loginService.getUserByEmail();
-            if (name == null) {
-                model.addAttribute("renovations", renovationRecordService.getRecordResultByUser(user));
-            } else if (name.trim().isEmpty()) {
-                model.addAttribute("renovations", renovationRecordService.getRecordResultByUser(user));
-            } else {
-                model.addAttribute("renovations", renovationRecordService.getRecordResultByName(user, name));
-            }
+            model.addAttribute("renovations", renovationRecordService.getRecordResultByName(user, searchQuery));
+            model.addAttribute("searchQuery", searchQuery);
             return "renovationsTemplate";
         } catch (IllegalArgumentException e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
@@ -71,7 +65,7 @@ public class RenovationController {
      * Gets the renovation creation form
      * @return thymeleaf createRenovationTemplate
      */
-    @GetMapping("/renovations/create")
+    @GetMapping("/create")
     public String record() {
         logger.info("GET /renovations/create");
         return "createRenovationTemplate";
@@ -96,7 +90,7 @@ public class RenovationController {
      *              with values being set to relevant parameters provided
      * @return thymeleaf createRenovationTemplate OR viewRenovationTemplate
      */
-    @PostMapping("/renovations/create")
+    @PostMapping("/create")
     public String submitRecord(@RequestParam(name="name") String name,
             @RequestParam(name = "description", required=false, defaultValue = "") String description,
             @RequestParam(name = "roomList", required = false) List <String> roomList,
@@ -114,29 +108,43 @@ public class RenovationController {
         }
 
         try { // save the record and go to the view page
-            RenovationRecord renovationRecord = new RenovationRecord(name, description, roomList);
-            renovationRecordService.addRenovationRecord(renovationRecord);
-            model.addAttribute("renovation", renovationRecord);
-            return "viewRenovation";
+            User user = loginService.getUserByEmail();
+            try {
+                RenovationRecord renovationRecord = new RenovationRecord(user, name, description, roomList);
+                renovationRecordService.addRenovationRecord(renovationRecord);
+                model.addAttribute("renovation", renovationRecord);
+                return "viewRenovation";
+            } catch (IllegalArgumentException e) {
+                logger.warn("Form submission error", e);
+                model.addAttribute("name", name);
+                model.addAttribute("description", description);
+                model.addAttribute("roomList", roomList);
+                model.addAttribute("errorMessage", "Invalid input: " + e.getMessage());
+                return "createRenovationTemplate";
+            }
         } catch (IllegalArgumentException e) {
-            logger.warn("Form submission error", e);
-            model.addAttribute("name", name);
-            model.addAttribute("description", description);
-            model.addAttribute("roomList", roomList);
-            model.addAttribute("errorMessage", "Invalid input: " + e.getMessage());
-            return "createRenovationTemplate";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
         }
     }
 
     /**
      * Deletes renovation record by its id, redirects back to my records page
      * @param id of the record to be deleted
-     * @return redirect to my records page
+     * @return response based on whether the record id exists, if the user doesn't have permission to delete the record, or
+     * if the deletion was successful
      */
-    @PostMapping("/delete-renovation")
-    public String deleteRecord(@RequestParam("id") Long id) {
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<Void> deleteRecord(@PathVariable Long id) {
+        logger.info("DELETE /renovations/");
+        RenovationRecord record = renovationRecordService.getRecordById(id);
+        if (record == null) {
+            return ResponseEntity.notFound().build();
+        }
+        if (!loginService.getUserByEmail().equals(record.getUser())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         renovationRecordService.removeRenovationRecord(id);
-        return "redirect:/renovations";
+        return ResponseEntity.noContent().build();
     }
 
     /**
@@ -146,10 +154,10 @@ public class RenovationController {
      * with values being set to relevant parameters provided
      * @return Thymeleaf editRenovationTemplate
      */
-    @GetMapping("/renovations/edit")
+    @GetMapping("/edit")
     public String editRenovation(@RequestParam(name = "id") Long id, Model model) {
         RenovationRecord renovationRecord = renovationRecordService.getRecordById(id);
-        if (renovationRecord == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        if (renovationRecord == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "This renovation does not exist");
         model.addAttribute("renovation", renovationRecord);
         model.addAttribute("name", renovationRecord.getName());
         return "editRenovationTemplate";
@@ -183,7 +191,7 @@ public class RenovationController {
      * @param model (map-like) representation of results to be used by thymeleaf
      * @return redirect to the view page for the edited record
      */
-    @PostMapping("/renovations/edit")
+    @PostMapping("/edit")
     public String submitRenovationEdit(@RequestParam(name = "id") Long id,
                                        @RequestParam(name="name", required = false) String name,
                                        @RequestParam(name = "description", required = false) String description,
@@ -215,7 +223,7 @@ public class RenovationController {
      * @param model (map-like) representation of results to be used by thymeleaf
      * @return redirect to viewRenovation page
      */
-    @GetMapping("/renovations/view")
+    @GetMapping("/view")
     public String viewRenovation(@RequestParam(name = "id") Long id, Model model) {
         RenovationRecord record = renovationRecordService.getRecordById(id);
         if (record == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "This renovation does not exist");
